+++ src/main/kotlin/com/github/victorrentea/livecoding/FrameworkDetector.kt
@@ -3,14 +3,13 @@ package com.github.victorrentea.livecoding
 import com.intellij.openapi.project.Project
 import com.intellij.psi.JavaPsiFacade
 import com.intellij.psi.PsiElement
-import com.intellij.psi.search.GlobalSearchScope
 
 object FrameworkDetector {
-    fun lombokIsPresent(project: Project, element: PsiElement) =
-        JavaPsiFacade.getInstance(project)
+    fun lombokIsPresent(element: PsiElement) =
+        JavaPsiFacade.getInstance(element.project)
             .findClass("lombok.RequiredArgsConstructor", element.resolveScope) != null
+    fun slf4jIsPresent(element: PsiElement) =
+        JavaPsiFacade.getInstance(element.project)
+            .findClass("org.slf4j.LoggerFactory", element.resolveScope) != null
 
-    fun visibleForTestingIsPresent(project: Project, element: PsiElement) =
-        JavaPsiFacade.getInstance(project)
-            .findClass("com.google.common.annotations.VisibleForTesting", element.resolveScope) != null
 }