+++ CHANGELOG.md
@@ -24,3 +24,4 @@
 - Support auto-adding static imports to fields , not only methods
 - Add more static imports :  "java.lang.System#currentTimeMillis",  "org.mockito.ArgumentMatchers#anyInt",   "java.util.concurrent.TimeUnit#MILLISECONDS",  "java.time.Duration#ofSeconds",
   "java.time.Duration#ofMillis"
+- Quickfix to define a Slf4J logger on the class (besides Lombok @Slf4j) on unresolved "log" token